requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
email-validator
